✨ Feat: Confirm 컴포넌트 로직과 스타일 분리
